Fix get_vars_from_sys_argv() in PytorchExperiment
Used to call .get on an argparse.Namespace object, which doesn't
exist. Now we access the desired attribute directly. The parser
will make the attributes default to None, but we can be sure
they exist, so no worries there (Alternative would have been
getattr with None default, which we don't need).